requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2
python-dotenv==1.0.1
faiss-cpu==1.9.0
sentence-transformers==3.3.1
httpx==0.27.2
requests==2.32.3
pandas==2.2.3
numpy==1.26.4
tavily-python==0.5.0
